% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{initialize,Candidate-method,}
\alias{print,Candidate-method}
\alias{show,Candidate-method,}
\title{A Candidate Object}
\description{
Objects of class \code{Candidate} are created by the \code{createCandidate} function
}
\details{
An object of the class 'Candidate' has the following slots:
\itemize{
 \item{name}{The name of a candidate}
 \item{delegatesWon}{The number of delegates won by the candidate}
 \item{party}{The name of the candidate's party}
 \item{delegatesNeeded}{The total number of delegates needed by the candidate. For 
                        Democratic candidates, it is 2383, and for Republican 
                        candidates, it is 1237}
 }
}
\author{
Taishi Muraoka: \email{tmuraoka@wustl.edu}
}

