% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Create an Object with class "Candidate"}
\usage{
createCandidate(name, delegatesWon, party, ...)
}
\arguments{
\item{name}{A character object with the name of the candidate}

\item{delegatesWon}{A numeric object with the number of delegates won by the 
candidate}

\item{party}{A character object with the name of the candidate's party}
}
\value{
An object of class "Candidate" containing
 \item{name}{The name of a candidate}
 \item{delegatesWon}{The number of delegates won by the candidate}
 \item{party}{The name of the candidate's party}
 \item{delegatesNeeded}{The total number of delegates needed by the candidate. For 
                        Democratic candidates, it is 2383, and for Republican 
                        candidates, it is 1237}
}
\description{
Create an instance with S4 class "Candidate"
}
\examples{
createCandidate("Obama", 150, "Democratic")
createCandidate("Bush", 400, "Republican")

}
\author{
Taishi Muraoka \email{tmuraoka@wustl.edu}
}
\seealso{
\code{\link{PropNeeded}}, \code{\link{createRace}}
}

