% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emba_calculate}
\alias{emba_calculate}
\title{Perform EBMA using C++}
\usage{
emba_calculate(x, y, weights, sd, limit)
}
\arguments{
\item{x}{numeric x values}

\item{y}{observed outcomes}

\item{weights}{weights of each model}

\item{sd}{deviation}

\item{limit}{change before model stops}
}
\description{
Perform EBMA using C++
}
