% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,Race-method}
\alias{plot,Race-method}
\title{Plot an Object with class "Race"}
\usage{
\S4method{plot}{Race}(x, y)
}
\arguments{
\item{x}{An object of class "Race"}

\item{y}{Anything}
}
\value{
Create a barplot that shows the number of delegates each candidate won and the number of delegates each candidate has to gain in order to win the nomination
}
\description{
Create a plot using with S4 class "Race"
}
\examples{
obama <- createCandidate("obama", 333, "Republican")
bush <- createCandidate("bush", 30, "Republican")
gore <- createCandidate("gore", 500, "Republican")
cand_list <- list(obama, bush, gore)
candidates <- createRace(cand_list)
plot(candidates, "ANYTHING")

}
\author{
Taishi Muraoka \email{tmuraoka@wustl.edu}
}
\seealso{
\code{\link{createRace}}
}

