% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded.R
\name{propNeeded}
\alias{propNeeded}
\alias{propNeeded,ANY-method}
\title{The Proportion of Remaining Delegates Required for Winning}
\usage{
propNeeded(candidate, remaining, ...)
}
\arguments{
\item{candidate}{An object of class "Candidate"}

\item{remaining}{The number of remaining deligates}
}
\value{
An object of subclass "Remaining" containing
 \item{name}{The name of a candidate}
 \item{delegatesWon}{The number of delegates won by the candidate}
 \item{party}{The name of the candidate's party}
 \item{delegatesNeeded}{The total number of delegates needed by the candidate. For 
                        Democratic candidates, it is 2383, and for Republican 
                        candidates, it is 1237}
 \item{remain_prop}{The proportion of delegates remaining needed for the candidate 
                    to win}
}
\description{
Calculate the proportion of delegates remaining needed for the candidate to win
}
\examples{
obama <- createCandidate("Obama", 333, "Democratic")
propNeeded(obama, 1500)

}
\author{
Taishi Muraoka: \email{tmuraoka@wustl.edu}
}
\seealso{
\code{\link{createCandidate}}
}

