% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race.R
\docType{class}
\name{Race-class}
\alias{Race-class}
\alias{initialize,Race-method}
\title{A Race Object}
\description{
Objects of class \code{Race} are created by the \code{createRace} function
}
\details{
An object of the class 'Race' has the following slots:
\itemize{
 \item{all_candidates}{The names of all candidates}
 \item{party}{The name of the candidates' party}
 \item{remaining}{The number of remaining delegates}
 }
}
\author{
Taishi Muraoka: \email{tmuraoka@wustl.edu}
}

