% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remaining.R
\docType{class}
\name{Remaining-class}
\alias{Remaining-class}
\alias{initialize,Remaining-method}
\title{A Remaining object}
\description{
Object of class \code{Remaining} as created by the \code{propNeeded} function
}
\details{
An object of the class `Squares' has the following slots:
\itemize{
 \item{name}{The name of a candidate}
 \item{delegatesWon}{The number of delegates won by the candidate}
 \item{party}{The name of the candidate's party}
 \item{delegatesNeeded}{The total number of delegates needed by the candidate. For 
                        Democratic candidates, it is 2383, and for Republican 
                        candidates, it is 1237}
 \item{The proportion of delegates remaining needed for the candidate to win}
 }
}
\author{
Taishi Muraoka: \email{tmuraoka@wustl.edu}
}

