% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{z}
\alias{z}
\title{predicted zvalues}
\usage{
z(x, y, weight, sd)
}
\arguments{
\item{x}{matrix of value}

\item{y}{actual values}

\item{weight}{weight for each model}

\item{sd}{deviations}
}
\description{
predicted zvalues
}
